New method show_container_item_context_menu
This metho implements the container functionality.
While this is a container context menu the idea is that it is controlled by the main GUI class  therefore it is implemented there.
However, it has to be callable from the container class therefore it has to be passed inside as a handle.